+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Released: Upcoming
 - Added filters for the list views
 - Big cleanup of the CRUDEntityDefinition constructor
 - Not required number fields not being entered by the user end up the database as NULL instead of as 0
+- null is a valid value when validating an int, float or reference which is not required in an entity
 
 ## 0.9.6
 Released: 2015-02-20